+++ include/embree4/rtcore_common.h
@@ -298,6 +298,8 @@ enum RTCFeatureFlags
     RTC_FEATURE_FLAG_USER_GEOMETRY_CALLBACK_IN_ARGUMENTS |
     RTC_FEATURE_FLAG_USER_GEOMETRY_CALLBACK_IN_GEOMETRY,
 
+  RTC_FEATURE_FLAG_32_BIT_RAY_MASK = 1 << 28,
+
   RTC_FEATURE_FLAG_ALL = 0xffffffff,
 };
 
@@ -465,7 +467,7 @@ inline decltype(F) rtcGetSYCLDeviceFunctionPointer(sycl::queue& queue)
 {
   sycl::buffer<cl_ulong> fptr_buf(1);
   {
-    auto fptr_acc = fptr_buf.get_access<sycl::access::mode::write>();
+    auto fptr_acc = fptr_buf.get_host_access(); // <sycl::access::mode::write>
     fptr_acc[0] = 0;
   }
 
@@ -477,7 +479,7 @@ inline decltype(F) rtcGetSYCLDeviceFunctionPointer(sycl::queue& queue)
     });
   queue.wait_and_throw();
 
-  auto fptr_acc = fptr_buf.get_access<sycl::access::mode::read>();
+  auto fptr_acc = fptr_buf.get_host_access(); // <sycl::access::mode::read>
   return (decltype(F)) fptr_acc[0];
 }