+++ kernels/rthwif/rthwif_embree.cpp
@@ -55,6 +55,9 @@ const constexpr uint32_t TRAV_LOOP_FEATURES =
   RTC_FEATURE_FLAG_TRIANGLE |   // filter function enforced for triangles and quads in this case
   RTC_FEATURE_FLAG_QUAD |
 #endif
+#if defined(EMBREE_RAY_MASK)
+  RTC_FEATURE_FLAG_32_BIT_RAY_MASK |
+#endif
   RTC_FEATURE_FLAG_MOTION_BLUR |
   RTC_FEATURE_FLAG_CURVES |
   RTC_FEATURE_FLAG_GRID |
@@ -265,7 +268,7 @@ bool intersect_instance(intel_ray_query_t& query, Ray& ray, Instance* instance, 
 template<typename Ray>
 bool intersect_primitive(intel_ray_query_t& query, Ray& ray, Scene* scenes[RTC_MAX_INSTANCE_LEVEL_COUNT+1], Geometry* geom, sycl::private_ptr<RayQueryContext> context, uint32_t geomID, uint32_t primID, const RTCFeatureFlags feature_mask)
 {
-#if defined(__SYCL_DEVICE_ONLY__)
+#if defined(EMBREE_SYCL_SUPPORT) && defined(__SYCL_DEVICE_ONLY__)
   bool filter = feature_mask & (RTC_FEATURE_FLAG_FILTER_FUNCTION_IN_ARGUMENTS | RTC_FEATURE_FLAG_FILTER_FUNCTION_IN_GEOMETRY);
   if (feature_mask & RTC_FEATURE_FLAG_MOTION_BLUR) {
     if (ray.time() < geom->time_range.lower || geom->time_range.upper < ray.time())
@@ -542,7 +545,9 @@ void trav_loop(intel_ray_query_t& query, Ray& ray, Scene* scenes[RTC_MAX_INSTANC
     Geometry* geom = scene->get(geomID);
 
     /* perform ray masking */
+#if defined(EMBREE_RAY_MASK)
     if (ray.mask & geom->mask)
+#endif
     {
       if (candidate == intel_candidate_type_procedural)
       {