+++ kernels/rthwif/rthwif_production_igc.h
@@ -4,8 +4,14 @@
 #pragma once
 
 #include <cstdint>
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+#pragma clang diagnostic ignored "-W#pragma-messages"
+
 #include <sycl/sycl.hpp>
 
+#pragma clang diagnostic pop
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wreturn-type-c-linkage"