+++ kernels/rthwif/testing/rthwif_cornell_box.cpp
@@ -15,6 +15,10 @@
 #include <vector>
 #include <iostream>
 
+#if !defined(_unused)
+#define _unused(x) ((void)(x))
+#endif
+
 void exception_handler(sycl::exception_list exceptions)
 {
   for (std::exception_ptr const& e : exceptions) {
@@ -112,6 +116,7 @@ void* alloc_accel_buffer(size_t bytes, sycl::device device, sycl::context contex
   void* ptr = nullptr;
   ze_result_t result = zeMemAllocShared(hContext,&device_desc,&host_desc,bytes,RTHWIF_ACCELERATION_STRUCTURE_ALIGNMENT,hDevice,&ptr);
   assert(result == ZE_RESULT_SUCCESS);
+  _unused(result);
   return ptr;
 }
 
@@ -120,6 +125,7 @@ void free_accel_buffer(void* ptr, sycl::context context)
   ze_context_handle_t hContext = sycl::get_native<sycl::backend::ext_oneapi_level_zero>(context);
   ze_result_t result = zeMemFree(hContext,ptr);
   assert(result == ZE_RESULT_SUCCESS);
+  _unused(result);
 }
 
 #define bvh_bytes 4992