+++ scripts/package_linux.sh
@@ -14,7 +14,7 @@ function check_symbols
     if [ ${#sym} -le 1 ]; then
         continue;
     fi;
-    version=(`echo $sym | sed 's/.*@@\(.*\)$/\1/g' | grep -E -o "[0-9]+"`)
+    version=(`echo $sym | cut -d"@" -f2 | grep -E -o "[0-9]+"`)
     if [ ${#version[@]} -ne 0 ]; then
       if [ ${#version[@]} -eq 1 ]; then version[1]=0; fi
       if [ ${#version[@]} -eq 2 ]; then version[2]=0; fi
@@ -51,10 +51,11 @@ EMBREE_SIGN_FILE=$5
 
 # create package
 make -j 16 preinstall
-#check_symbols lib${EMBREE_LIBRARY_NAME}.so GLIBC 2 4 0
-check_symbols lib${EMBREE_LIBRARY_NAME}.so GLIBC 2 14 0    # GCC 4.8
-check_symbols lib${EMBREE_LIBRARY_NAME}.so GLIBCXX 3 4 14
-check_symbols lib${EMBREE_LIBRARY_NAME}.so CXXABI 1 3 0
+
+check_symbols lib${EMBREE_LIBRARY_NAME}.so GLIBC 2 28 0
+check_symbols lib${EMBREE_LIBRARY_NAME}.so GLIBCXX 3 4 22
+check_symbols lib${EMBREE_LIBRARY_NAME}.so CXXABI 1 3 11
+
 make -j 16 package
 
 if [ "$EMBREE_ZIP_MODE" == "OFF" ]; then
@@ -63,7 +64,7 @@ if [ "$EMBREE_ZIP_MODE" == "OFF" ]; then
   if [ $# -eq 5 ]; then
     ${EMBREE_SIGN_FILE} embree${EMBREE_VERSION_MAJOR}-*-${EMBREE_VERSION}-*.rpm
   fi
-    
+
   # create TGZ of RPMs
   embree_tgz=embree-${EMBREE_VERSION}.x86_64.rpm.tar.gz
   tar czf ${embree_tgz} embree${EMBREE_VERSION_MAJOR}-*-${EMBREE_VERSION}-*.rpm