+++ tutorials/embree_tests/kernels/kernels_tests.cpp
@@ -1,5 +1,9 @@
 // Copyright 2009-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
+#if !defined(_CRT_SECURE_NO_WARNINGS)
+#define _CRT_SECURE_NO_WARNINGS
+#endif
+
 #include "common/common_tests.cpp"
 #include "geometry/geometry_tests.cpp"